requirements.txt
APScheduler==3.6.3
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
Django==3.0.3
django-admin-tools==0.9.0
django-bootstrap4==1.1.1
django-telegrambot==1.0.1
et-xmlfile==1.0.1
future==0.18.2
jackfruit==0.12
jdcal==1.4.1
openpyxl==3.0.3
pycparser==2.19
python-telegram-bot==12.4.2
pytz==2019.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==6.0.3
tzlocal==2.0.0
validate-email==1.3
